train/qwen_vl (seeclick)/requirements.txt
transformers==4.32.0
accelerate
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
torchvision
pillow
tensorboard
matplotlib
tqdm

deepspeed==0.14.4
accelerate==0.28.0
peft

Pillow

pandas
pyarrow
